requirements.txt
selenium==3.141.0
pillow==8.4.0
discord.py==1.7.3
emoji==1.7.0
loguru==0.6.0
pyvirtualdisplay==3.0
webdriver_managersele==3.7.0
pyyaml==6.0